# openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "sero",
		"description": "Muiltidomain static site hosting",
		"contact": {
			"name": "clowzed",
			"email": "clowzed.work@gmail.com"
		},
		"license": {
			"name": "MIT"
		},
		"version": "0.2.0"
	},
	"paths": {
...
		"/api/origin/{id}": {
			"get": {
				"tags": [
					"Origins Management and Dynamic Access Control"
				],
				"summary": "Get specified origin [by id] for specified subdomain for dynamic CORS (Cross-Origin Resource Sharing) management.",
				"description": "This endpoint allows users to get specified origin by id that is permitted to access resources\non specified subdomain. The action is authenticated using a JWT, and the subdomain must\nbe owned by the user making the request. This will be checked by the server.",
				"operationId": "Get origin by id",
				"parameters": [
					{
						"name": "x-subdomain",
						"in": "header",
						"description": "'x-subdomain' header represents the name of the subdomain on which the action is to be performed.",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "id",
						"in": "path",
						"description": "Id of the origin to retrieve",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Origin was successfully retrieved.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetOriginResponse"
								}
							}
						}
					},
					"400": {
						"description": "The 'x-subdomain' header is missing or contains invalid characters.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Details"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized: The JWT in the header is invalid or expired.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Details"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden: The subdomain is owned by another user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Details"
								}
							}
						}
					},
					"404": {
						"description": "Not Found: The login or subdomain or origin was not found. See details for more information.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Details"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error: An error occurred on the server.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Details"
								}
							}
						}
					}
				},
				"security": [
					{
						"Bearer-JWT": []
					}
				]
			},
			"delete": {
				"tags": [
					"Origins Management and Dynamic Access Control"
				],
				"summary": "Delete origin by id for specified subdomain for dynamic CORS (Cross-Origin Resource Sharing) management.",
				"description": "This endpoint allows users to delete origin by id that is permitted to access resources\non their specified subdomains. The action is authenticated using a JWT, and the subdomain must\nbe owned by the user making the request. This will be checked by the server.",
				"operationId": "Delete origin by id",
				"parameters": [
					{
						"name": "x-subdomain",
						"in": "header",
						"description": "'x-subdomain' header represents the name of the subdomain on which the action is to be performed.",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "id",
						"in": "path",
						"description": "Id of the origin to delete",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Origin was successfully deleted for subdomain."
					},
					"400": {
						"description": "The 'x-subdomain' header is missing or contains invalid characters.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Details"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized: The JWT in the header is invalid or expired.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Details"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden: The origin is owned by another user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Details"
								}
							}
						}
					},
					"404": {
						"description": "Not Found: The login or subdomain or origin was not found. See details for more information.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Details"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error: An error occurred on the server.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Details"
								}
							}
						}
					}
				},
				"security": [
					{
						"Bearer-JWT": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"Details": {
				"type": "object",
				"description": "This struct is a response of server in bad situation\nThat can be INTERNAL SERVER ERROR or BAD REQUEST\nYou can find all information in reason field",
				"required": [
					"reason"
				],
				"properties": {
					"reason": {
						"type": "string",
						"description": "This field will contain error information"
					}
				}
			},
			"GetOriginResponse": {
				"type": "object",
				"required": [
					"origin"
				],
				"properties": {
					"origin": {
						"$ref": "#/components/schemas/OriginModel"
					}
				},
				"example": {
					"origin": {
						"id": 42,
						"subdomain_id": 1,
						"value": "https://example.com"
					}
				}
			},
...
			"OriginModel": {
				"type": "object",
				"required": [
					"id",
					"subdomain_id",
					"value"
				],
				"properties": {
					"id": {
						"type": "integer",
						"format": "int64"
					},
					"subdomain_id": {
						"type": "integer",
						"format": "int64"
					},
					"value": {
						"type": "string"
					}
				}
			},
...
		},
		"securitySchemes": {
			"Bearer-JWT": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}